+++ components/molecule/breadcrumb/src/index.js
@@ -1,36 +1,61 @@
-/* eslint-disable react/prop-types */
-import {useState} from 'react'
 import PropTypes from 'prop-types'
-import Chevronright from '@s-ui/react-icons/lib/Chevronright'
 
-import {breadcrumbClassName} from './settings.js'
+import useControlledState from '@s-ui/react-hooks/lib/useControlledState/index.js'
+import ChevronRight from '@s-ui/react-icons/lib/Chevronright'
+
+import {breadcrumbClassName, BASE_CLASS, isFunction} from './settings.js'
 
 const BreadcrumbBasic = ({
   items,
   icon,
-  linkFactory: Link,
-  isScrollable = false
+  linkFactory: Link = ({to, href, className, children}) => (
+    <a href={to || href} className={className}>
+      {children}
+    </a>
+  ),
+  isScrollable = false,
+  isExpanded,
+  defaultIsExpanded = false,
+  onExpand,
+  onCollapse,
+  onClick
 }) => {
-  const [isExpanded, setIsExpanded] = useState(false)
-  const expandBreadcrumb = () => setIsExpanded(true)
+  const [isExpandedState, setIsExpandedState] = useControlledState(
+    isExpanded,
+    defaultIsExpanded
+  )
+  const handleClick = event => {
+    setIsExpandedState(!isExpandedState)
+    isFunction(onClick) && onClick(event, {value: !isExpandedState})
+    if (isExpandedState) {
+      isFunction(onCollapse) && onCollapse(event, {value: false})
+    } else {
+      isFunction(onExpand) && onExpand(event, {value: true})
+    }
+  }
 
-  const IconAngle = icon || Chevronright
+  const IconAngle = icon || ChevronRight
   const numItems = items.length - 1
 
   return (
     <nav aria-label="breadcrumb" role="navigation">
-      <div className={breadcrumbClassName({isExpanded, isScrollable})}>
-        <button onClick={expandBreadcrumb} className="sui-BreadcrumbBasic-btn">
+      <div
+        className={breadcrumbClassName({
+          isExpanded: isExpandedState,
+          isScrollable
+        })}
+      >
+        <button onClick={handleClick} className={`${BASE_CLASS}-btn`}>
           ...
         </button>
-        <ul className="sui-BreadcrumbBasic-list">
+        <ul className={`${BASE_CLASS}-list`}>
           {items.map(({url, label}, index) => (
-            <li className="sui-BreadcrumbBasic-listItem" key={index}>
+            <li className={`${BASE_CLASS}-listItem`} key={index}>
               {index !== 0 && index <= numItems && (
-                <IconAngle svgClass="sui-BreadcrumbBasic-icon" />
+                <IconAngle svgClass={`${BASE_CLASS}-icon`} />
               )}
               {url ? (
-                <Link to={url} href={url} className="sui-BreadcrumbBasic-link">
+                <Link to={url} href={url} className={`${BASE_CLASS}-link`}>
                   {label}
                 </Link>
               ) : (
@@ -73,15 +98,21 @@ BreadcrumbBasic.propTypes = {
   /**
    * Boolean that allows us to show the items with a horizontal scroll
    */
-  isScrollable: PropTypes.bool
-}
-
-BreadcrumbBasic.defaultProps = {
-  linkFactory: ({to, href, className, children}) => (
-    <a href={to || href} className={className}>
-      {children}
-    </a>
-  )
+  isScrollable: PropTypes.bool,
+  /**
+   * The controlled value of the expanded l&f of the component
+   */
+  isExpanded: PropTypes.bool,
+  /**
+   * The initial value of the expanded l&f of the component
+   */
+  defaultIsExpanded: PropTypes.bool,
+  /** expand handler **/
+  onExpand: PropTypes.func,
+  /** collapse handler **/
+  onCollapse: PropTypes.func,
+  /** click handler (expand or collapse) **/
+  onClick: PropTypes.func
 }
 
 export default BreadcrumbBasic