+++ components/molecule/breadcrumb/src/settings.js
@@ -1,9 +1,11 @@
 import cx from 'classnames'
 
-const BASE_CLASS = 'sui-BreadcrumbBasic'
+export const BASE_CLASS = 'sui-BreadcrumbBasic'
 
 export const breadcrumbClassName = ({isExpanded, isScrollable}) =>
   cx(BASE_CLASS, {
     'is-expanded': isExpanded,
     'is-scrollable': isScrollable
   })
+
+export const isFunction = fn => typeof fn === 'function'